+++ packager/media/base/playready_key_source.h
@@ -16,28 +16,36 @@
 namespace shaka {
 namespace media {
 
-/// A key source that uses playready for encryption.
+/// A key source that uses PlayReady for encryption.
 class PlayReadyKeySource : public KeySource {
  public:
   /// Creates a new PlayReadyKeySource from the given packaging information.
   /// @param server_url PlayReady packaging server url.
-  /// @param proteciton_systems_flags is the flags indicating which PSSH should
+  /// @param protection_systems_flags is the flags indicating which PSSH should
   ///        be included.
+  /// @param protection_scheme is the Protection Scheme to be used for
+  ///        encryption. It needs to be signalled in Widevine PSSH. This
+  ///        argument can be ignored if Widevine PSSH is not generated.
   PlayReadyKeySource(const std::string& server_url,
-                     int protection_scheme_flags);
+                     int protection_systems_flags,
+                     FourCC protection_scheme);
   /// Creates a new PlayReadyKeySource from the given packaging information.
   /// @param server_url PlayReady packaging server url.
   /// @param client_cert_file absolute path to a client certificate.
   /// @param client_cert_private_key_file absolute path to the private file
   ///     for the client certificate.
   /// @param client_cert_private_key_password password for the private key.
-  /// @param proteciton_systems_flags is the flags indicating which PSSH should
+  /// @param protection_systems_flags is the flags indicating which PSSH should
   ///        be included.
+  /// @param protection_scheme is the Protection Scheme to be used for
+  ///        encryption. It needs to be signalled in Widevine PSSH. This
+  ///        argument can be ignored if Widevine PSSH is not generated.
   PlayReadyKeySource(const std::string& server_url,
                      const std::string& client_cert_file,
                      const std::string& client_cert_private_key_file,
                      const std::string& client_cert_private_key_password,
-                     int protection_scheme_flags);
+                     int protection_systems_flags,
+                     FourCC protection_scheme);
   ~PlayReadyKeySource() override;
 
   /// @name KeySource implementation overrides.
@@ -68,7 +76,6 @@ class PlayReadyKeySource : public KeySource {
  private:
   Status GetKeyInternal();
   Status GetCryptoPeriodKeyInternal();
-  explicit PlayReadyKeySource(std::unique_ptr<EncryptionKey> key);
 
   std::unique_ptr<EncryptionKey> encryption_key_;
   std::string server_url_;